examples/requirements.txt
aws-cdk-lib==2.114.1
aws-cdk.aws-scheduler-alpha==2.111.0a0
aws-cdk.aws-scheduler-targets-alpha==2.111.0a0
constructs>=10.0.0,<11.0.0
